routes and utils separated into own folders & files
