--- app/main.wsgi
@@ -1,4 +1,0 @@
-from main import app
-
-if __name__ == "__main__":
-    app.run()
